Improve testing in linera-views
