+++ linera-service/tests/end_to_end_tests.rs
@@ -12,8 +12,8 @@ use linera_service::{
     cli_wrappers::{Database, LocalNetwork, Network},
     util,
 };
-use linera_views::common::get_table_name;
-use std::{path::PathBuf, time::Duration};
+use linera_views::test_utils::get_table_name;
+use std::time::Duration;
 
 #[tokio::test]
 async fn test_resolve_binary() {
